Moved toString method to GroupFigure, as a templatemethod, which can be overwritten
